platform/android: replace eventfd with full bidirectional control channel
In the future it will come in handy to have a full bidirectional
channel between the service Java world and the JS thread.
Implement that using Unix domain sockets (with a filesystem path
because socketpair() is ugly to use from Android) with simple
JSON based marshalling.
Two methods are supported: foo (echo a number) and stop (stop
the engine). Thanks to the latter now stopping a service is
reliable.