Refactor the engine, fill in databases, introduce apps
The previous IFTTT concept was too limited, let's replace it
with the concept of "apps". Apps are simple objects with
a start() and a stop() that can be instantiated given a node.js
module on the server (or builtin) using arbitrary stored parameters.

Also introduce an "ifttt" kind of app, which implements exactly
what Rulepedia did (and tries to be compatible with it), with
all drawbacks and limitations.