Add rudimentary UI support
The frontend can now obtain a reference to an app, and ask it to
produce some UI, which can be a raw HTML string or a reference
to a Jade view.

For now, only the server platform supports it, because it is easy
there. Cloud needs a way to arbitrate the different running engine
instances, phone needs to emulate the web server in the webkit UI
it uses internally (or just use cloud)