Use syncdb for apps
Use a syncdb that hosts state and location, so that apps that should
run on a given tier can mark that information in the db.
A stub AppLocationManager is introduced, which will contain the
algorithm to choose a location for each app, and the engine is
updated to handle apps coming and going.