Introduce first-class ThingEngine instances
By representing running ThingEngine instances as devices in
the DB, we get automatic free syncing of configuration
parameters: when a new pairing is established, both devices
immediately receive knowledge of the third and can configure
themselves.
The use cases that are unblocked by this are:
- log in to the cloud from the server UI
  pair the phone with the QR code
  --> phone figures out the cloud ID
- log in to the cloud from the server UI
  log in to the cloud from the phone UI
  --> phone figures out the server address and port

(Originally, this work had a completely different purpose,
namely, instantiating different channels for UPnP on phone
vs UPnP on server. This turned out to be a cool side quest)