transparent-rpc: new module for seamless RPC in node
To be split out to an independent project and live on its own.

Provides a RPC API similar in spirit to Java RMI, with transparent
construction of proxy objects based on class metadata. Works over
any stream transport and it uses a format similar to JSON RPC
(but not identical).
Like in Java, objects are sent by value if they don't have RPC
metadata, and by reference otherwise. Methods can be called on
objects sent by reference.