platform/cloud: enable transparent-rpc
Clean up the child process message handling from tier_connections
(which now uses a private platform feature like in the server case).
A wrapper is added around process.send / process.on('message') to
get a Stream-like API that rpc.Socket requires.

Then make it so that the engine process will send a reference to
the engine object to the frontend process.