engine: add new app executor based on the new language
After discussion with Jiwon, this is the new language and the
new semantics we'll be used.

Config pairing is moved to a module of the engine, and we change
the schema to separate code and state, and have appdb
instantiate the app executor directly.

Different kind of apps are no longer supported, they were a bad
idea anyway. Test apps have been reimplemented using the new
syntax.

UI code in platform/server/frontend and platform/cloud was
updated to have only one kind of app (and greatly simplified
in the process)