DeviceDiscovery: clean up the code and try to make sense of it...
We were copying the chat example too closely, and it was hard
to understand what the protocol was meant to be. Now at least
it has one purpose: we own a name, we export an object on said
name, and at times (which for now is never) we emit a signal
on that object.
Plain and simple.