Add rudimentary virtual assistant system
The virtual assistant is implemented by a couple "Dialog" classes
in engine/assistant_manager.js, and it's proxied through RPC
to the Omlet implementation running in the control process of the
cloud platform, which is associated to the Sabrina account.
Each engine is associated to one Omlet feed when Sabrina is enabled,
and messages exchanged on that feed are routed to the appropriate
engine.