platform/cloud: run engines in a sandbox
Use a modified version of the xdg-app sandbox to run each engine
in a limited sandbox where the only writable file system is /app
(which is bind mounted to the current directory of the engine),
/srv and /home are completely invisible, /usr, /var and /run are
readonly, /sys and /dev are limited, and the no new privileges
flag is turned on.
Pid namespaces are employed to limit ptrace attack surface.