Implement queries
Queries are one-off triggers, or action/commands that produce data.
They are implemented by codegenning some appropriate ThingTalk,
then loading it and letting the core code take care of it.
This is somewhat an overkill (compared to just calling invokeQuery
on the appropriate channel, like ActionDialog does), but it will
be useful in the future if additional language capabilities are
to be enabled.