Nerf @sabrina
Now that we send anything we don't understand to bing (and thus
we "understand" everything), @sabrina.listen is completely broken.
@sabrina.onpicture is somewhat less broken but still not really
useful.
@sabrina.say/@sabrina.picture are still working but one should
use the new formatting abilities of @$notify for richer content.

Unfortunately, we cannot kill it completely or existing setups
will start spewing errors because they have the device in the
db (as unfortunately it's not a transient device). But we can
nerf all its channels until it's no longer introducting complexity.