Implement new location format
The bulk of this patch is the ability to resolve values based
on the context of the user, for example the current location,
prior to invoking ThingTalk.
The patch is made ugly by the fact that the current subdialog
mechanism does not work if the handle() function of a subdialog
is asynchronous. To fix that, handleGeneric needs to return
a promise, which causes a bit of code churn.