Implement slot filling of feeds
I want to use @omlet.send to test, but I need to slot fill feeds
to do that.

Unfortunately this does not work because we cannot codegen a feed
constant in ThingTalk. And even if we did, we need someone to
open the feed when it's passed to @omlet.send().
Feeds will need rethinking soon.