DeviceView: fix race conditions related to subviews
DeviceView.values() was completely bogus when dealing with subviews
(because they were completey ignored), but also addOne/removeOne
were creating races because they forced the value to be a promise
and the add to happen at the next mainloop iteration.
We don't need any of that, so we can copy the code from ObjectSet.Simple
and de-promiseify it.