Make formatting local and synchronous
Formatting should never involve any I/O: if any data is needed, it
should be part of the event in the first place.
It also should not need any special capability beyond standard JS
primitives.

Therefore we can run it locally using a non-proxy version of the
channel (which we have anyway because we loaded the device).