Refactor how we download devices
Now we always download the full manifest (using a new version of the
download API), and we use that to guide how the rest of the device
module should be downloaded.
This means that we find when a cached module is out of date by downloading
the manifest again and checking the version (which we now do automatically
once a week), and it opens the doors to more loaders for declaratively
defined thingpedia interfaces, as well as other formats than "thingpedia.v1".