Remove graphdb
A datastore for the user is something we certainly want, but it
should be in a separate module, probably using a full database.
The database will only work on the home server/cloud server
installation, not in Android, and should be accessible as a standard
Thingpedia device and not a core feature.