Add a new way to write dialogs
Dialogs are now asynchronous functions which can use .ask() to
retrieve data from the user. A StateMachineDialog provides integration
with the existing code.
To simplify further, the async functions are written as generators
(using babel's async-to-generator helper), until node grows support
for async/await in ES2017.

For now, RuleDialog and DiscoveryDialog (as well as their dependencies)
have been converted.