Redo how we present output of immediate queries
Instead of pushing the results to the general notification queue,
which will be drained whenever and does not let the API user
know when the commands are done, have a second, separate, output
queue which the API user is supposed to drain to get outputs and
errors from commands.

To keep things uniform, we introduce an abstraction of two delegates, OutputQueue
and ConversationOutput. ExecWrapper was changed to dispatch to
the delegate received upon construction, and all code was changed
to report results and errors through ExecWrapper.