sqlite: rework how locking and transactions works
We can't rely on acquiring and releasing the Database object to
ensure transaction isolation if we want to use indexedb-js, so
rework withTransaction() to roll its own transaction-level queuing
and make withClient() semi-synchronous.
This means that withTransaction() acquires an exclusive lock over
the database and withClient() acquires no lock at all, which is
half-broken, but it should be fine for our use cases.