Slightly optimize channel states
Because all channels have state now, we might end up with a lot
of unused rows in the state table. To avoid unused disk writes,
avoid hitting the database if the state did not change, and to
avoid wasting space, assume that a missing entry in the state
table corresponds to the default empty state.