Remove flaky test
The test was based on the assumption that the server would use
a nearest-neighbor classifier with a bag-of-words encoding (hence
an exact subsequence would always match), but currently the
server uses the encoder network from the regular parser to compute
the sentence embedding and that means the assumption does not
hold.

We should not have tests that depend on model quality, as they
break too often.