device_choice: fail with a generic message when a builtin device is missing
Builtin devices (gnome, home, phone) are only useful if you're using
Almond on the correct platform. For that reason, chances are that
a command issued for that device on any other platform is an incorrect
parse rather than an explicit command against something that is not
allowed.

Indicate this fact by failing with a generic "I did not understand"
error rather than a cryptic and weird-looking "You don't have a Home".