messaging/device_manager: add support for managing more than one device at once
Very soon almond-cloud will grow a simple communication layer
to allow users on the same server to communicate without setting
up a matrix account.

At the same time, we don't want to lock users in to a more limited
messaging platform if they have already configured matrix, and we
don't want matrix to conflict with the platform messaging.

So instead we create a virtual platform, that is the union of
all configured messaging account. This means that a user can
use matrix, omlet or the almond builtin one freely, and will then
communicate using the appropriate account.

When resolving an identity (email/phone), we use the first messaging
interface that recognizes the identity (with priority to the platform
one).

Includes tests, that finally actually test matrix in some form
(and could make us marginally more confident to upgrade matrix-js-sdk
to a more recent version)