slot_filling/contact_search: try platform data to resolve usernames
If platform data is provided with the current intent, use it to
resolve usernames to/from principals.

The goal here is that bridges such as the Slack bridge would provide
a platform data object mapping the username (as used in the command,
which is potentially an opaque id like in Slack) to both the display
value to use in the confirmation, and to either the communication
principal directly, or an identity that can further be resolved
to a communication principal.