Make use of statistics from the platform, if provided
Instead of using the engine's statistics module, which stores
statistics per-user on disk and is neither efficient nor useful,
use the platform's implementation.

On almond-cloud, this will be hooked up to Prometheus and the
rest of the metric infrastructure we have.

On other platforms, there will be no implementation and we'll
use a dummy module that does nothing. This is not a regression
from before, given that before we were recording those numbers
with no hope of ever collecting them.