Remove statistics module
It's a lot of untested code, of dubious use, because the API we actually
need is just one method (hit()), and because we don't have a way to
collect this data. Instead, platforms should provide platform-specific
implementation of a statistics collector, if they care. Otherwise,
we won't collect anything.

Worth mentioning that the only user of statistics was in almond-dialog-agent,
not here.