Training: run without outputdir, and handle being killed
If the job is killed in between processes, job.child will be null.
When that happens, stop the job at the first chance instead of
going further.

Also, support not passing an outputdir. almond-cloud/training will
not want to do a cp of the saved-model, preferring instead to use
a symlink as that's faster and more atomic.