Introduce contextual templates
Contextual templates involve a new type of non-terminal, which
is declared of type "context" ahead of time.
Different context non-terminals can be declared, for contexts
of different types.

The generator tool takes an existing dataset to be the source
of potential contexts, and then runs the expansion grammar on
each one.