Refactor cloud sync manager to support multiple clients to the same server
Assign a unique ID to each device owned by the user, and separate
the "tier" (whether we're a phone/server/cloud - which affects which
way we connect and how we initiate sync) from the "address" (which
is both tier and the unique ID).
This will allow syncing multiple phones against the same cloud account,
and incidentally also allows syncing a phone against multiple home
servers.

Includes tests, and includes a bunch of related bug fixes