FileThingpediaClient: switch to loading ThingTalk instead of JSON
ThingTalk class files are easier to read and edit by hand, which
makes it easy to apply changes such as adding or removing devices
in a local snapshot. They are also easily greppable, which
makes it easy to figure out which devices are part of an experiment
and which are excluded.

Entities cannot be represented in ThingTalk, so split them
into a separate file.