Add progress estimation to the sentence generator
Estimating progress is hard, because it depends on the (future)
behavior of semantic functions.
This commit introduces a series of heuristics based on our
adaptive estimation of each rule productivity. These heuristic
result in a monotonic progress bar, but not in a very stable one.
Luckily, it gets more stable as time progresses / higher depth,
which is also where most of the time is spent.

Fixes #49