ReplaceParameters: use proper parsing & typechecking to identify replaceable parameters
The previous code would choke on filters of a join, because it would
assign all parameters to the second function.
The code was ok when dealing with normal synthetic output, because
we never carry over a parameter from the first function, but it
would fail on contextual synthetic, which does.

The new code assigns parameters to the right function correctly,
the same way almond-dialog-agent and other code does.

Note that the test cases did not change, which makes me confident
the new code is at least no more broken than the old code.