Mark all contextual commands as confident
In the current implementation (until we implement stack frames)
contextual commands hit the reset path and start ruleDialog from scratch,
but we don't want to confirm them again because that would be inconvenient.
Assume that understanding contextual commands is easier, and treat them
as confident.

Also, increase the timeout for contexts to 5 minutes.