lib/parallelize: implement back pressure
If all workers are busy (not reading from the stream), stop accepting
data from the input stream. This ensures that the input stream is advanced
at a controlled rate compared to the output stream, and allows measuring the
progress of the overall operation by measuring the progress of the
input stream.