Use a generic message for "no configured device"
Because users don't know what is supported and what is not supported,
we can get into "no configured devices" both when they try something
that is supported but they haven't configured, and when they try
something unsupported and the parser gets confused. With an ambiguous
message we can kick them to the skill configuration page and they
see everything.