replace_parameters: fix accessing arg if there is no primitive
If we're replacing inside the filter of a join, there is no
current primitive (because can come from either primitive), so
"slot.primitive" is null.
Instead, we can access the argument directly from the slot.
Use either the soon-to-be-added public "arg" property of the slot,
or the private "_arg" property, and fall back to trying the prim
anyway if we have one.