Rework how constants are handled in sentence-generation
At "inference time", we want to extract the available constants
from the current context, instead of generating numbered constants.

To do so, the runtime tracks which rules generates constants,
and asks the target language module to extract all values from
the context. As a side effect, this removes an import of thingtalk
from the generation runtime (which was misplaced), removes a HACK
(which was unnecessary) and removes all differences in common-constants.genie
and load-thingpedia.js between thingtalk and dlgthingtalk