Add $agent templates for a message after the end of the conversation
In Genie, turns go "agent -> user", because that's how Multiwoz
was written, and it's how the neural network should see the context
to generate things.

In Almond, turns go "user -> agent", because that's how virtual
assistants work.

We would like turns to be paired always, so add a dialogue act to
the agent that can be invoked after the end of the dialogue.