Add Genie-based dialogue policy
The Genie based dialogue policy runs the Genie templates in
"inference mode" to actually chose the agent action and the next
sentence.

This way, we don't have to write the agent policy twice, once
in the form of templates to simulate the dialogue and produce
training data, and once in the form of the actual policy here.

The downside is that the policy is fairly slow to run. I think
the templates can be optimized to prune more intelligently though.