Implement related questions
"Related questions" are questions that are introduced in the
second turn and use a different function than the first question,
but carry over all the filters. They can be used to implement:

U: Hello, I'm looking for a Chinese restaurant.
A: We have many Chinese restaurants. ....
U: What are restaurants that I booked recently?
A: I found X ...

This is part of the 294S homework, and all the parts specific
to related questions (not bugfixes or changes to load-thingpedia)
will be stripped out.