Template & annotation tweaking
- Make description not filterable
We actually do not use description at all - we consider a question not answerable
if we don't have a specific property for it when collecting eval and test.
We are setting ourselves into trouble introducing description filters.

- Make brand not filterable
We never use it, and in our data we only have one brand

- Add templates with table canonical behind names

- More annotaitons for recipes and products