Fix GPT-2 ranker used to rank annotations
First, the output logits should be shifted one token to right
before calculating the probability because in autoregressive
language models like GPT-2, each logit is a probability distribution
for the `next` token.

Second, we changed from calculating probabilities to calculating
perplexities. Therefore, lower numbers are more "natural". This also
results in bigger values for outputs (~10-500 instead of ~0.0001-0.1)
which might be easier to interpret.