Add a tokenizer for Chinese
This is mostly to test how well we do for ideographic languages.
There is no statistical tokenization: we split all ideographs
into their own tokens, except those that we recognize (dates, times, etc.).

It is quite shorter than English or Italian because the rules
for numbers are dates are much, much simpler. And we can even
understand colloquial times!