Import AssistantDispatcher and speech code from almond-server
Move history handling to the Conversation class, and import
a generic version of the AssistantDispatcher, so we don't have
to copy it in all platforms.

Also import the speech handling code (reacting to a wake-word,
streaming voice to the STT, calling the TTS), which is also
shared between multiple platforms. The actual sound handling
code is still abstracted behind the platform, so the same
code might work on Windows/Mac/Android/iOS.