load-thingpedia: allow loading only a limited subset of Thingpedia
In inference mode, we don't want to dump the whole Thingpedia onto
the user's device and try to make sense of all of it. If we do so,
we end with a lot of confusion when generating the agent's reply,
and more often than not end up generating nothing, especially for
the most challenging dialogue acts, like agent confirmations.

Instead, we can extract the relevant devices from the current
state, and initialize the grammar only for those functions. The
same grammar is used from one turn to the next when the devices
are the same, and transparently reinitialized when a new function
is added.