device-choice: remove support for choosing among multiple devices
The device ID is not part of the state representation in NN syntax,
so it is lost when the neural network predicts the state at the
next turn (replacing the current state). Hence, we can't choose
a device manually, so let's always choose the first device.

This is temporary, until the device ID is actually added as a
first class object in NN-syntax and device selection is added
to the state machine (which we will need for IoT).