simulator: always generate small numbers with some probability
During batch dialogue generation, only small numbers can be used
in agent sentences, because the agent picks from a fixed pool
of constants (same as the user). So if we don't generate small
numbers we end up with no agent utterances for certain parameters,
which in turn prune out important contexts.