templates: simplify how ID entities are identified
Instead of loading the query while processing entities.json,
record the fact that an entity is an ID entity when processing
the associated query.

This avoids attempting to load a ton of Thingpedia devices due
to random unrelated entities in entities.json, which slows
down the agent quite a bit when talking to almond-dev's Thingpedia.