Restore entity constants of ID type at the first turn
We need them to pick up "songs by taylor swift", where "taylor
swift" is a entity constant as value to the "artist" filter.

At the same time, we want "book at mcdonalds" or "play despacito"
to be "query + action", so we reject them if the use the constant
directly, and also make sure there are no "id==" filters
(otherwise one cannot search for a "mcdonalds in palo alto")

I tried converting the action to query + action in the initial
request dialogue act, but it doesn't work: the sentence will
use a GENERIC_ENTITY_ token and the program will use a string,
which will break things. We might be able to do a postprocessing
that fixes both sentence and program but it feels like a hack.