Restore compound commands that are executed for all query results
When the user issues "play all songs released by taylor swift this year"
we want to convert that to a compound command and automatically
queue all songs, rather than showing the top three songs and then
letting the user choose.
Similarly, when the user says "play delicate by taylor swift" we
want to directly choose the right song and play it, instead of
asking the user confirmation.

Because displaying the result of the query is some sort of
confirmation, this is now controlled by the #[confirm] annotation,
which is turned into a 3-state enum (but "true" and "false" still work).
If `#[confirm=enum(auto)]`, then we generate compound commands;
`#[confirm=enum(display_result)]`, we split into a query and action
statement.