Fix action results in the real executor
The "output type" of an action has "action/" as a prefix, and we
need to compute that correctly.
Also, we need to propagate input parameters to the result, in case
the action does not do that (it is not required to), because we
want to use the actual input parameter in the reply from the agent.