Add an option to not delete inactive conversations
We want to delete conversations that are inactive (no activity
for more than 5 minutes and the connection is closed) in Web Almond
to reclaim memory, but we don't want to delete the "main" conversation
in almond-server/almond-gnome (the conversation backed by speech)
so the user can scroll back and see the history.

At the same time, we still want the inactivity timeout to kick in
so the context is reset, and the microphone is closed if the agent
is asking a question.