Remove sound support from the test assistant (#257)
"genie assistant" is meant to be a quick way to test a new model
interactively. Supporting speech is nice, but it has heavy,
platform-specific dependencies (including the wake word stuff).

It was useful to have the code (even though it was not actually
working in practice) as guidance on how to enable almond-server,
which has real sound. Now that almond-server is up to date,
let's remove this.