evaluation: apply thingtalk normalizations more consistently
Ensure that .optimize() is called on the reference prediction
and on the output of the network, to apply semantic normalization
(e.g. removing redundant computations or projection), as well
as toNN() to apply syntactic normalization (sorting of parameter
names and filters)

As it happens, this requires quite a bit of refactoring, but
after that we're applying exactly the same code paths between
evaluate-server and evaluate-dialog, which seems a good thing.