Adapt auto-annotate-multiwoz.js for 2.2
This essentially copies over most of the original script, into the
new one (which processes the annotations for MultiWOZ 2.2). The script
for 2.2 is not working fully, as far as I can tell.

This is a temporary branch for multiple reasons, all centering around
the fact that this code is very dirty. It is only being pushed remotely
to allow for collaboration, but it's temporary and will be deleted soon.