Include incomplete values in SGD database
If an action errors out, we never get to see all values of the action.
But we will still need all values of the relevant entity. So we add
those to the db with `null` values, to be filled out in postprocessing.

Also output the incorrect predicted code in wrong_syntax cases during
eval.