Implement better handling of non-canonicals
Now we use canonicalMap to keep track of the canonical versions of all
non-canonical values. This is important in situations where the
canonical value is implicitly the same as some previous value mentioned
in some earlier point in the dialogue. Usually this happens when a slot
is carried over from one domain to another, and the slot in the new
domain is never given a canonical value (but by matching the
non-canonical form from the previous domain, we can find a canonical
value).