Add #[default] annotation for optional input parameters
The annotation provides explicit, training-time knowledge of a
default value. It can be used to tap into the user's context
with values like $context.location.here. It can also be used
to remove ambiguity.

At training time, if a parameter is explicitly given in the
natural language, it is removed from the code (as if it had
no effect). That way, it can be freely removed from paraphrasing,
because it is implicit.
At execution time, the parameter is added again if unspecified.