Add a tool to subsample a thingpedia.tt file
When generating dialogues, we run 6 generation in parallel. We
can be a lot more efficient if these 6 run on different subsets
of Thingpedia, as each process sees a smaller Thingpedia and
can generate more varied datasets on it.
This reduces the amount of wasted work caused by a large Thingpedia,
and reduces the amount of duplicated sentences across processes.

The downside is that we lose some cross-domain transitions,
but the neural network can usually still pick those up fairly
well, as long as some cross-domain transitions are included.