speech: remove automatic trigger
Auto-trigger is meant to do voice-activity-detection so that
any speech can continue the conversation while Almond is asking
a question, without issuing the wake-word again.

Unfortunately, it does not work well. Without echo-cancellation,
it gets triggered by Almond speaking or playing music. With
echo-cancellation, there is enough feedback noise from the ec
algorithm that the VAD algorithm triggers constantly.
If we want VAD in the future, we should use a better library.

Fixes #288