Fix terminating conversations without going into an infinite loop
To avoid looping, we need to distinguish the 'cancel' state
(to which the agent replies "would you like anything else")
from the 'end' state (to which the agent says "good bye" and
resets).

We have this distinction in the state machine and during training,
but both state a triggered by essentially the same phrase, and
the cancel state is far more common than the end state, because
it can appear at any time. As a result, the parser overwhelmingly
prefers to generate 'cancel' instead of 'end' for those states.

To obviate this problem, apply a prediction time heuristic
to override the state.

Fixes #298