templates: handle empty action results correctly
If the action produced no result, even though it is expected to
produce one result, use an empty search template to show a
message to the user. To do that, we need to let the empty search
templates operate on the context even though the current function
is not the query function.

Fixes #312