augment: fix single-device augmentation
The issue was the the dialogue policy token, which starts with @
like Thingpedia devices. The fix is to use the same code as
the evaluation path, where we had a similar problem before.

Fixed that, make sure we don't apply the augmentation after the
first turn (it doesn't make sense) and drastically reduce the
default expansion.