Download data for one property at a time instead of one item(entity) at a time
In Schema.org, we download data one page at a time - which normally contains information for
all properties of a single entity.
This approach is no longer efficient for Wikidata due to some limitations of the SPARQL endpoint:
we have to get a list of entities first, and then request values for each property one query at a
time - this is very slow, yet still get very sparse information as lots of entities don't have
value for most of the properties.

Instead, we just query example values for each property directly, without generate a data.json:
we will not use the data.json anyway, since we will translate thingtalk to SPARQL and query against
wikidata.org directly.