templates: fix applying the default priority boost
The canonical form indicated in the "default" key should receive
a priority boost and be preferred when making agent sentences.
But for that to work, we need to check the renamed annotation,
because when we check we use the old name of the POS, and the
category has the new name.