Schema.org: Add the option to override parameter name
The override will not change the actual paramter name in the manifest, rather
it's used to replace the parameter name as the base to generate canonicals.

Also update canonical override loader: iteratively removing prefixes when
dealing with compound types - always try the full name first, and gradually
removing prefixes until a match is found.