Add infrastructure for typechecking templates
To typecheck templates, we need to compile them into TypeScript
instead of directly to JS, and then invoke the TypeScript compiler
on it.
We can invoke the typescript compiler programmatically,
but if we do so we don't really get typechecking, because TS
needs to load the modules and compile a whole bunch of files
to typecheck correctly.
So instead we add a new tool, "genie compile-template", that
converts a .genie file to .genie.ts file in the same folder.
The .genie.ts file can then be compiled as regular TypeScript,
with the usual typechecking based on the imports.

This commit does not yet add type annotations to the Genie
syntax. That's the next step, and will require more work as
annotations need to be picked up at the point where a non-terminal
is defined.