Fix running tests in nyc
In nyc we can compile typescript on the file (to instrument it
for code coverage) but when we run on the basic starter we're
pointing to the precompiled templates, not the sources. To fix
that, try loading the precompiled js first, and fail back to
on-demand compilation only if that does not succeed.