Remove Chinese templates, multiwoz templates for ACL
With the migration to TypeScript, and soon to happen changes
in ThingTalk, these are unlikely to work, and I don't want to
spend the effort to keep them building, so just delete them
for now.

For translation, templates will be specified in a different way,
and for ACL, people should checkout the exact commit used in
the paper to reproduce the results.