Add type annotations to the sentence generation code
This is actually very messy, because the sentence generation code
will operate on different types depending on what target language
is chosen (thingtalk or multidst), and will also load templates
with their own types. But in the end it types (with enough "any"
here and there) and it forced me to rewrite some ugly code where
we were mixing ReservoirSampler and array. It also caught some
edge case bug that would never trigger with our templates but
were wrong nevertheless.