Remove nltk wordnet dependency and use a custom common word list instead
The custom words is based on /usr/share/dict/linux.words filtered by nltk wordnet synset.
The file is preprocessed and stored in https://almond-static.stanford.edu/test-data/common-words.txt

The file is optional to run bert anntoator: if the file is missing, we will skip the step of
filtering uncommon words.