sentence-generator: introduce a new, faster sampling strategy
Our previous sampling strategy consisted of enumerating, for each
template, all possible expansions of that template, and then
sampling a portion of those expansions to call the semantic function
with. This is not efficient. Instead, we can sample directly
one candidate expansion without enumerating, if we are careful
with setting up the probabilities correctly.