Fix requotable augmentation
If we're running in requotable mode (which is the default), we should
not apply the pluralize transformation.

This was caught by the multiwoz starter tests: they do not have
an eval set so they run split-train-eval, which requires the dataset
to be requotable. It started failing now because we changed the
sampling strategy enough to generate a different dataset.