Switch to ThingTalk 2.0
This is a large commit, containing almost entirely mechanical changes:
we first change the package.json to refer to the ThingTalk 2.0
branches of ThingTalk and Thingpedia, and then fix the code until
it typechecks.

The templates are also changed to operate in terms of the new AST
types exclusively, rather than using the compatibility interfaces
in ThingTalk. This simplifies tracking the types, because we don't
need to keep track of what code is old and what code is new, at the
cost of a bigger diff.

All code that deals with ThingTalk syntax is converted to use
internal helpers (except code that parses Thingpedia). The helpers
will try to parse both new and old syntax. They always serialize
new syntax though, except in "genie server" when we invoke
ThingTalk's compat code based on the advertised client version.