Automatically construct an exact match model for developer devices
Import the exact matcher code (including tests) from almond-cloud,
and use it to automatically build an exact match model for all
the devices in the developer directory, when one is specified.
The exact match model is cached until the manifests are modified
again.
The exact match is applied both for the local parser client
(--nlu-server file:///) and for the remote parser client, overriding
the remote reply.