Change synthesis so contexts non-terms optionally carry a phrase
Instead of mapping context non-terminals always to the empty
string, make it possible to map them to a non-empty string that
is concatenated to the rest of the template as usual. The restriction
that context non-terminals must be first is also removed, and
multiple context non-terminals can appear in the same template.

This allows using top-down deterministic logic (ie, the same logic
used by the describe module and the old text formatter) to construct
some portions of the agent reply. This should make the agent reply
code much faster at inference time, because we don't need to enumerate as
much (indeed, we should not enumerate at all in the most common
cases), and should avoid hitting cases were we produce no reply at
all for certain thingtalk states, because we don't have a bottom-up
template for that state.